440: Split vote checks into separate files
